requirements.txt
python-mnist
matplotlib==2.0.2
numpy==1.13.3
Pillow==4.3.0
